+++ src/prod_cons/framed.rs
@@ -1,27 +1,95 @@
+//! Framed byte queue interfaces
+//!
+//! Useful for sending data that has coherent frames, e.g. network packets
+
 use core::{
     marker::PhantomData,
     ops::{Deref, DerefMut},
     ptr::NonNull,
 };
 
-use crate::{
-    queue::BBQueue,
-    traits::{
-        bbqhdl::BbqHandle,
-        coordination::{Coord, ReadGrantError, WriteGrantError},
-        notifier::{AsyncNotifier, Notifier},
-        storage::Storage,
-    },
+use crate::traits::{
+    bbqhdl::BbqHandle,
+    coordination::{Coord, ReadGrantError, WriteGrantError},
+    notifier::{AsyncNotifier, Notifier},
+    storage::Storage,
 };
 
+/// A trait that can be used as the "header" for separating framed storage.
+///
+/// Framed interfaces use a `u16` by default, which only requires two bytes
+/// for the header, with the limitation that the largest grant allowed is
+/// 64KiB at a time. You can also use a `usize` allowing the maximum platform
+/// available size.
+///
+/// You should not have to implement this trait.
+///
 /// # Safety
 ///
 /// Do it right
 pub unsafe trait LenHeader: Into<usize> + Copy + Ord {
+    /// Should be `[u8; size_of::<Self>()]`
     type Bytes;
+    /// Convert Self into Self::Bytes, in little endian order
     fn to_le_bytes(&self) -> Self::Bytes;
+    /// Convert Self::Bytes (which is in little endian order) to Self.
     fn from_le_bytes(by: Self::Bytes) -> Self;
 }
+
+/// A producer handle that can be used to write framed chunks
+pub struct FramedProducer<Q, H = u16>
+where
+    Q: BbqHandle,
+    H: LenHeader,
+{
+    pub(crate) bbq: Q::Target,
+    pub(crate) pd: PhantomData<H>,
+}
+
+/// A consumer handle that can be used to read framed chunks
+pub struct FramedConsumer<Q, H = u16>
+where
+    Q: BbqHandle,
+    H: LenHeader,
+{
+    pub(crate) bbq: Q::Target,
+    pub(crate) pd: PhantomData<H>,
+}
+
+/// A writing grant into the storage buffer
+///
+/// Grants implement Deref/DerefMut to access the contained storage.
+#[must_use = "Write Grants must be committed to be effective"]
+pub struct FramedGrantW<Q, H = u16>
+where
+    Q: BbqHandle,
+    H: LenHeader,
+{
+    bbq: Q::Target,
+    base_ptr: NonNull<u8>,
+    hdr: H,
+}
+
+/// A reading grant into the storage buffer
+///
+/// Grants implement Deref/DerefMut to access the contained storage.
+///
+/// Write access is provided for read grants in case it is necessary to mutate
+/// the storage in-place for decoding.
+#[must_use = "Read Grants must be released to free space"]
+pub struct FramedGrantR<Q, H = u16>
+where
+    Q: BbqHandle,
+    H: LenHeader,
+{
+    bbq: Q::Target,
+    body_ptr: NonNull<u8>,
+    hdr: H,
+}
+
+// ---- impls ----
+
+// ---- impl LenHeader ----
 
 unsafe impl LenHeader for u16 {
     type Bytes = [u8; 2];
@@ -36,6 +104,7 @@ unsafe impl LenHeader for u16 {
         u16::from_le_bytes(by)
     }
 }
+
 unsafe impl LenHeader for usize {
     type Bytes = [u8; core::mem::size_of::<usize>()];
 
@@ -50,53 +119,18 @@ unsafe impl LenHeader for usize {
     }
 }
 
-impl<S: Storage, C: Coord, N: Notifier> BBQueue<S, C, N> {
-    pub fn framed_producer(&self) -> FramedProducer<&'_ Self> {
-        FramedProducer {
-            bbq: self.bbq_ref(),
-            pd: PhantomData,
-        }
-    }
-
-    pub fn framed_consumer(&self) -> FramedConsumer<&'_ Self> {
-        FramedConsumer {
-            bbq: self.bbq_ref(),
-            pd: PhantomData,
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl<S: Storage, C: Coord, N: Notifier> crate::queue::ArcBBQueue<S, C, N> {
-    pub fn framed_producer(&self) -> FramedProducer<std::sync::Arc<BBQueue<S, C, N>>> {
-        FramedProducer {
-            bbq: self.0.bbq_ref(),
-            pd: PhantomData,
-        }
-    }
-
-    pub fn framed_consumer(&self) -> FramedConsumer<std::sync::Arc<BBQueue<S, C, N>>> {
-        FramedConsumer {
-            bbq: self.0.bbq_ref(),
-            pd: PhantomData,
-        }
-    }
-}
-
-pub struct FramedProducer<Q, H = u16>
-where
-    Q: BbqHandle,
-    H: LenHeader,
-{
-    pub(crate) bbq: Q::Target,
-    pub(crate) pd: PhantomData<H>,
-}
+// ---- impl FramedProducer ----
 
 impl<Q, H> FramedProducer<Q, H>
 where
     Q: BbqHandle,
     H: LenHeader,
 {
+    /// Attempt to obtain a write grant of the given (max) size
+    ///
+    /// The returned grant can be used to write up to `sz` bytes, though
+    /// a smaller size may be committed. Dropping the grant without calling
+    /// commit means that no data will be made visible to the consumer.
     pub fn grant(&self, sz: H) -> Result<FramedGrantW<Q, H>, WriteGrantError> {
         let (ptr, cap) = self.bbq.sto.ptr_len();
         let needed = sz.into() + core::mem::size_of::<H>();
@@ -121,25 +155,32 @@ where
     Q::Notifier: AsyncNotifier,
     H: LenHeader,
 {
+    /// Wait for the given write grant to become available
+    ///
+    /// If `sz` is larger than the storage buffer, this method will never
+    /// return.
+    ///
+    /// The returned grant can be used to write up to `sz` bytes, though
+    /// a smaller size may be committed. Dropping the grant without calling
+    /// commit means that no data will be made visible to the consumer.
     pub async fn wait_grant(&self, sz: H) -> FramedGrantW<Q, H> {
         self.bbq.not.wait_for_not_full(|| self.grant(sz).ok()).await
     }
 }
 
-pub struct FramedConsumer<Q, H = u16>
-where
-    Q: BbqHandle,
-    H: LenHeader,
-{
-    pub(crate) bbq: Q::Target,
-    pub(crate) pd: PhantomData<H>,
-}
+// ---- impl FramedConsumer ----
 
 impl<Q, H> FramedConsumer<Q, H>
 where
     Q: BbqHandle,
     H: LenHeader,
 {
+    /// Attempt to receive a single frame
+    ///
+    /// The FramedConsumer has no control over the size of the read grant,
+    /// we see whatever size was written by the FramedProducer.
+    ///
+    /// The returned grant must be released to free the space in the buffer.
     pub fn read(&self) -> Result<FramedGrantR<Q, H>, ReadGrantError> {
         let (ptr, _cap) = self.bbq.sto.ptr_len();
         let (offset, grant_len) = self.bbq.cor.read()?;
@@ -194,23 +235,43 @@ where
     }
 }
 
-pub struct FramedGrantW<Q, H = u16>
+// ---- impl FramedGrantW ----
+
+impl<Q, H> FramedGrantW<Q, H>
 where
     Q: BbqHandle,
     H: LenHeader,
 {
-    bbq: Q::Target,
-    base_ptr: NonNull<u8>,
-    hdr: H,
-}
+    /// Commit `used` bytes of the grant to be visible.
+    ///
+    /// If `used` is greater than the `sz` used to create this grant, the
+    /// amount will be clamped to `sz`.
+    pub fn commit(self, used: H) {
+        let (_ptr, cap) = self.bbq.sto.ptr_len();
+        let hdrlen: usize = const { core::mem::size_of::<H>() };
+        let grant_len = hdrlen + self.hdr.into();
+        let clamp_hdr = self.hdr.min(used);
+        let used_len: usize = hdrlen + clamp_hdr.into();
 
-unsafe impl<Q, H> Send for FramedGrantW<Q, H>
-where
-    Q: BbqHandle,
-    Q::Target: Send,
-    H: LenHeader + Send
-{
+        unsafe {
+            self.base_ptr
+                .cast::<H>()
+                .as_ptr()
+                .write_unaligned(clamp_hdr);
+        }
 
+        self.bbq.cor.commit_inner(cap, grant_len, used_len);
+        self.bbq.not.wake_one_consumer();
+        core::mem::forget(self);
+    }
+
+    /// Aborts the grant, making no frame available to the consumer
+    ///
+    /// Can be used to silence "must_use" errors.
+    pub fn abort(self) {
+        // The default behavior is to abort - do nothing, let the
+        // drop impl run
+    }
 }
 
 impl<Q, H> Deref for FramedGrantW<Q, H>
@@ -259,53 +320,39 @@ where
     }
 }
 
-impl<Q, H> FramedGrantW<Q, H>
-where
-    Q: BbqHandle,
-    H: LenHeader,
-{
-    pub fn commit(self, used: H) {
-        let (_ptr, cap) = self.bbq.sto.ptr_len();
-        let hdrlen: usize = const { core::mem::size_of::<H>() };
-        let grant_len = hdrlen + self.hdr.into();
-        let clamp_hdr = self.hdr.min(used);
-        let used_len: usize = hdrlen + clamp_hdr.into();
-
-        unsafe {
-            self.base_ptr
-                .cast::<H>()
-                .as_ptr()
-                .write_unaligned(clamp_hdr);
-        }
-
-        self.bbq.cor.commit_inner(cap, grant_len, used_len);
-        self.bbq.not.wake_one_consumer();
-        core::mem::forget(self);
-    }
-
-    pub fn abort(self) {
-        // The default behavior is to abort - do nothing, let the
-        // drop impl run
-    }
-}
-
-pub struct FramedGrantR<Q, H = u16>
-where
-    Q: BbqHandle,
-    H: LenHeader,
-{
-    bbq: Q::Target,
-    body_ptr: NonNull<u8>,
-    hdr: H,
-}
-
-unsafe impl<Q, H> Send for FramedGrantR<Q, H>
+unsafe impl<Q, H> Send for FramedGrantW<Q, H>
 where
     Q: BbqHandle,
     Q::Target: Send,
-    H: LenHeader + Send
+    H: LenHeader + Send,
 {
+}
 
+// ---- impl FramedGrantR ----
+
+impl<Q, H> FramedGrantR<Q, H>
+where
+    Q: BbqHandle,
+    H: LenHeader,
+{
+    /// Release the entire read grant
+    ///
+    /// It is not possible to partially release a framed read grant.
+    pub fn release(self) {
+        let len: usize = self.hdr.into();
+        let hdrlen: usize = const { core::mem::size_of::<H>() };
+        let used = len + hdrlen;
+        self.bbq.cor.release_inner(used);
+        self.bbq.not.wake_one_producer();
+        core::mem::forget(self);
+    }
+
+    /// Drop the grant WITHOUT releasing the message from the queue.
+    ///
+    /// The next call to read will observe the same packet again.
+    pub fn keep(self) {
+        // Default behavior is "keep"
+    }
 }
 
 impl<Q, H> Deref for FramedGrantR<Q, H>
@@ -343,21 +390,10 @@ where
     }
 }
 
-impl<Q, H> FramedGrantR<Q, H>
+unsafe impl<Q, H> Send for FramedGrantR<Q, H>
 where
     Q: BbqHandle,
-    H: LenHeader,
+    Q::Target: Send,
+    H: LenHeader + Send,
 {
-    pub fn release(self) {
-        let len: usize = self.hdr.into();
-        let hdrlen: usize = const { core::mem::size_of::<H>() };
-        let used = len + hdrlen;
-        self.bbq.cor.release_inner(used);
-        self.bbq.not.wake_one_producer();
-        core::mem::forget(self);
-    }
-
-    pub fn keep(self) {
-        // Default behavior is "keep"
-    }
 }
